tests/requirements_2x.txt
mock
pylint
vcrpy
